requirements.txt
beautifulsoup4==4.12.3
httpx==0.27.2
Pillow==10.4.0
